+++ src/main.rs
@@ -1,4 +1,4 @@
-mod lib;
+use benders::mutations::*;
 
 fn main() {
     println!("Hello world!");
